Create integration test with and without using docker
