+++ .travis.yml
@@ -4,6 +4,9 @@ jdk:
   - oraclejdk8
   - oraclejdk9
 
+services:
+ - docker
+
 # cache settings
 before_cache:
   - rm -f  $HOME/.gradle/caches/modules-2/modules-2.lock
@@ -12,6 +15,9 @@ cache:
   directories:
     - $HOME/.gradle/caches/
     - $HOME/.gradle/wrapper/
+
+install:
+ - docker pull mongo
 
 script:
 - ./tutorial/gradlew --no-daemon -b tutorial/build.gradle test
