+++ tutorial/build.gradle
@@ -17,6 +17,7 @@ dependencies {
     testImplementation 'org.mockito:mockito-core:2.18.0'
     testImplementation 'com.github.fakemongo:fongo:1.6.5'
     testImplementation 'ch.qos.logback:logback-classic:1.1.1'
+    testImplementation 'org.testcontainers:database-commons:1.7.2'
     implementation 'org.mongodb:mongo-java-driver:2.14.3'
     implementation 'org.jongo:jongo:1.3.0'
 }
