+++ tutorial/src/test/java/attsw/exam/example/core/MongoRepositoryIT.java
@@ -1,0 +1,21 @@
+package attsw.exam.example.core;
+
+import java.net.UnknownHostException;
+
+import org.junit.ClassRule;
+import org.testcontainers.containers.GenericContainer;
+
+import com.mongodb.MongoClient;
+
+public class MongoRepositoryIT extends AbstractMongoRepositoryTest {
+
+	@SuppressWarnings("rawtypes")
+	@ClassRule
+	public static GenericContainer mongo = new GenericContainer("mongo:latest").withExposedPorts(27017);
+
+	@Override
+	protected MongoClient extractMongoClient() throws UnknownHostException {
+		return new MongoClient(mongo.getContainerIpAddress(), mongo.getMappedPort(27017));
+	}
+
+}
